requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
boto3==1.9.50
botocore==1.12.50
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
elasticsearch==6.3.1
idna==2.7
jmespath==0.9.3
lxml==4.2.5
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
PyPDF2==1.26.0
pytest==4.0.1
python-dateutil==2.7.5
python-docx==0.8.7
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
